django_backend/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
Pillow==10.4.0
cloudinary==1.41.0
python-dotenv==1.0.1